Crate spiritnet_runtime

The KILT runtime. This can be compiled with #[no_std], ready for Wasm.

Re-exports

pub use public_credentials;

Modules

api

xcm_config

Macros

add_benchmarks

Calls add_benchmark with all configs from define_benchmarks and passes the first two parameters on.

list_benchmarks

Calls list_benchmark with all configs from define_benchmarks and passes the first two parameters on.

Structs

AutoUnlockBound

Burn

ExistentialDeposit

InflationInfo

InstantAllowed

MaxApprovals

MaxAuthorities

MaxClaims

MaxFreezes

MaxLocks

MaxProposalWeight

MaxProposals

MaxReserves

MaxScheduledPerBlock

MaxVotes

MaximumSchedulerWeight

MinimumPeriod

NoPreimagePostponement

OriginPrivilegeCmp

Used the compare the privilege of an origin inside the scheduler.

PalletInfo

Provides an implementation of PalletInfo to provide information about the pallet setup in the runtime.

PreImageHoldReason

ProposalBond

ProposalBondMinimum

ReservedDmpWeight

ReservedXcmpWeight

Runtime

RuntimeApi

RuntimeApiImpl

Implements all runtime apis for the client side.

RuntimeGenesisConfig

RuntimeOrigin

The runtime origin type representing the origin of a call.

SS58Prefix

SessionKeys

SpendPeriod

TreasuryAccount

UncleGenerations

UsableBalance

Version

Enums

OriginCaller

ProxyType

The type used to represent the kinds of proxying allowed.

RuntimeCall

RuntimeError

RuntimeEvent

RuntimeFreezeReason

A reason for placing a freeze on funds.

RuntimeHoldReason

A reason for placing a hold on funds.

RuntimeLockId

An identifier for each lock placed on funds.

RuntimeSlashReason

A reason for slashing funds.

RuntimeTask

An aggregation of all Task enums across all pallets included in the current runtime.

Constants

VERSION

This runtime version.

WASM_BINARY

WASM_BINARY_BLOATY

Traits

BuildStorage

Complex storage builder stuff.

Functions

native_version

The version information used to identify this runtime when compiled natively.

Type Aliases

Address

The address format for describing accounts.

AllPalletsWithSystem

All pallets included in the runtime as a nested tuple of types.

AllPalletsWithoutSystem

All pallets included in the runtime as a nested tuple of types. Excludes the System pallet.

Attestation

Aura

AuraConfig

AuraExt

AuraExtConfig

Authorship

Balances

BalancesConfig

Block

Block header type as expected by this runtime. Block type as expected by this runtime.

BlockId

BlockId type as expected by this runtime.

CheckedExtrinsic

Extrinsic type that has already been checked.

Council

CouncilConfig

Ctype

CumulusXcm

Delegation

Democracy

DemocracyConfig

DepositStorage

Did

DidLookup

DidLookupConfig

DipProvider

DmpQueue

Executive

Executive: handles dispatch to the various modules.

GenesisConfigDeprecated

Indices

Inflation

MessageQueue

Migration

Multisig

ParachainInfo

ParachainInfoConfig

ParachainStaking

ParachainStakingConfig

ParachainSystem

ParachainSystemConfig

PolkadotXcm

PolkadotXcmConfig

Preimage

Proxy

PublicCredentials

Scheduler

Session

SessionConfig

SignedBlock

A Block signed with a Justification

SignedExtra

The SignedExtension to the basic transaction logic.

System

SystemConfig

TechnicalCommittee

TechnicalCommitteeConfig

TechnicalMembership

TechnicalMembershipConfig

Timestamp

Tips

TipsMembership

TipsMembershipConfig

TransactionPayment

Treasury

TreasuryConfig

UncheckedExtrinsic

Unchecked extrinsic type as expected by this runtime.

Utility

Vesting

VestingConfig

Web3Names

XcmpQueue